add KDialog token provider
